+++ config/decoder.go
@@ -125,7 +125,9 @@ func (d *decoder) sequence(childKey string, value reflect.Value) error {
 
 	// start looking for child values.
 	elementType := derefType(valueType).Elem()
-	childKey += _separator
+	if childKey != "" {
+		childKey += _separator
+	}
 
 	for ai := 0; ; ai++ {
 		arrayKey := childKey + strconv.Itoa(ai)
@@ -164,7 +166,9 @@ func (d *decoder) array(childKey string, value reflect.Value) error {
 
 	// start looking for child values.
 	elementType := derefType(valueType).Elem()
-	childKey += _separator
+	if childKey != "" {
+		childKey += _separator
+	}
 
 	for ai := 0; ai < value.Len(); ai++ {
 		arrayKey := childKey + strconv.Itoa(ai)
@@ -205,9 +209,18 @@ func (d *decoder) mapping(childKey string, value reflect.Value, def string) erro
 	// child yamlNode parsed from yaml file is of type map[interface{}]interface{}
 	// type casting here makes sure that we are iterating over a parsed map.
 	if v, ok := val.(map[interface{}]interface{}); ok {
-		childKey += _separator
+		if childKey != "" {
+			childKey += _separator
+		}
+
 		for key := range v {
 			mapKey := childKey + fmt.Sprintf("%v", key)
+
+			// Skip empty keys, because lookups in provider can toggle infinite loops/spill internals.
+			if mapKey == "" {
+				continue
+			}
+
 			itemValue := reflect.New(valueType.Elem()).Elem()
 
 			// Try to unmarshal value and save it in the map.