+++ config/yaml_benchmark_test.go
@@ -84,7 +84,7 @@ func BenchmarkYAMLPopulateStruct(b *testing.B) {
 	b.ResetTimer()
 
 	for n := 0; n < b.N; n++ {
-		p.Get("api.credentials").PopulateStruct(c)
+		p.Get("api.credentials").Populate(c)
 	}
 }
 
@@ -105,7 +105,7 @@ func BenchmarkYAMLPopulateStructNested(b *testing.B) {
 	b.ResetTimer()
 
 	for n := 0; n < b.N; n++ {
-		p.Get("api").PopulateStruct(s)
+		p.Get("api").Populate(s)
 	}
 }
 
@@ -128,7 +128,7 @@ func BenchmarkYAMLPopulateStructNestedMultipleFiles(b *testing.B) {
 	b.ResetTimer()
 
 	for n := 0; n < b.N; n++ {
-		p.Get("api").PopulateStruct(s)
+		p.Get("api").Populate(s)
 	}
 }
 
@@ -146,7 +146,7 @@ func BenchmarkYAMLPopulateNestedTextUnmarshaler(b *testing.B) {
 	b.ResetTimer()
 
 	for n := 0; n < b.N; n++ {
-		p.Get(Root).PopulateStruct(s)
+		p.Get(Root).Populate(s)
 	}
 }
 
@@ -161,7 +161,7 @@ encoderConfig:
 	b.ResetTimer()
 
 	for n := 0; n < b.N; n++ {
-		if err := p.Get(Root).PopulateStruct(cfg); err != nil {
+		if err := p.Get(Root).Populate(cfg); err != nil {
 			b.Error(err)
 		}
 	}